test: split fake-service construction and inspection coverage
