+++ pkg/services/factory_sessions/internal/execution/fake_service_runtime_test.go
@@ -90,10 +90,6 @@ func (portableRecordingTestWriter) Write(path string, value recordings.PortableR
 	return os.WriteFile(path, append(data, '\n'), 0o600)
 }
 
-func int64Ptr(value int64) *int64 {
-	return &value
-}
-
 func seedRuntimeSessionWithRunningDispatch(
 	service *JavaScriptRuntimeService,
 	sessionID, dispatchID, label string,
@@ -389,275 +385,6 @@ func TestProjectResultRead_DefaultsToFinalMode(t *testing.T) {
 	}
 	if projected.ResultStatus != ResultStatusFinal {
 		t.Fatalf("status = %q, want FINAL", projected.ResultStatus)
-	}
-}
-
-// pkgmaintcheck:ignore-cyclomatic-complexity this helper regression keeps lifecycle mutation assertions together on one scenario.
-func TestFakeService_InternalLifecycleHelpers(t *testing.T) {
-	t.Parallel()
-	state := &fakeSessionState{
-		session: SessionReadResult{
-			SessionID: "dur-sess-1",
-			Status:    LifecycleStatusAwaitingApproval,
-		},
-		result: ResultReadResult{
-			SessionID:     "dur-sess-1",
-			SessionStatus: LifecycleStatusAwaitingApproval,
-			ResultStatus:  ResultStatusNotReady,
-		},
-		dispatches: []DispatchSummary{{ID: "disp-1", Status: DispatchStatusFailed, Attempt: 1}},
-		dispatchDetails: map[string]DispatchDetail{
-			"disp-1": {DispatchSummary: DispatchSummary{ID: "disp-1", Status: DispatchStatusFailed, Attempt: 1}},
-		},
-	}
-
-	if err := validateLifecycleControlRequest(LifecycleControlApprove, ControlRequest{}, ApproveRequest{}, RetryDispatchRequest{}, InterruptDispatchRequest{}); err != nil {
-		t.Fatalf("approve validation: %v", err)
-	}
-	if err := validateLifecycleControlRequest(LifecycleControlRetryDispatch, ControlRequest{}, ApproveRequest{}, RetryDispatchRequest{}, InterruptDispatchRequest{}); err == nil {
-		t.Fatal("retry without dispatch id should fail validation")
-	}
-	if err := validateLifecycleControlRequest(LifecycleControlInterruptDispatch, ControlRequest{}, ApproveRequest{}, RetryDispatchRequest{}, InterruptDispatchRequest{}); err == nil {
-		t.Fatal("interrupt without dispatch id should fail validation")
-	}
-	if err := validateLifecycleControlRequest(LifecycleControlPause, ControlRequest{}, ApproveRequest{}, RetryDispatchRequest{}, InterruptDispatchRequest{}); err != nil {
-		t.Fatalf("pause validation: %v", err)
-	}
-
-	accepted := lifecycleControlResultFromState(state, "dur-sess-1", LifecycleControlApprove, LifecycleControlOutcomeAccepted, RetryDispatchRequest{}, InterruptDispatchRequest{})
-	if accepted.Session == nil || accepted.Session.Status != LifecycleStatusAwaitingApproval {
-		t.Fatalf("accepted lifecycle control result = %#v", accepted)
-	}
-	noop := lifecycleControlResultFromState(state, "dur-sess-1", LifecycleControlPause, LifecycleControlOutcomeNoOp, RetryDispatchRequest{}, InterruptDispatchRequest{})
-	if noop.Session == nil {
-		t.Fatalf("noop lifecycle control result = %#v", noop)
-	}
-	retry := lifecycleControlResultFromState(state, "dur-sess-1", LifecycleControlRetryDispatch, LifecycleControlOutcomeAccepted, RetryDispatchRequest{DispatchID: "disp-1"}, InterruptDispatchRequest{})
-	if retry.DispatchID != "disp-1" || retry.RetryDispatchID != "disp-1" {
-		t.Fatalf("retry lifecycle control result = %#v", retry)
-	}
-	interrupt := lifecycleControlResultFromState(state, "dur-sess-1", LifecycleControlInterruptDispatch, LifecycleControlOutcomeAccepted, RetryDispatchRequest{}, InterruptDispatchRequest{DispatchID: "disp-1"})
-	if interrupt.DispatchID != "disp-1" {
-		t.Fatalf("interrupt lifecycle control result = %#v", interrupt)
-	}
-
-	service := &FakeService{}
-	service.mutateSessionForControl(state, LifecycleControlApprove, RetryDispatchRequest{}, InterruptDispatchRequest{})
-	if state.session.Status != LifecycleStatusRunning {
-		t.Fatalf("approve mutate status = %q, want RUNNING", state.session.Status)
-	}
-	state.session.Status = LifecycleStatusFailed
-	state.result.SessionStatus = LifecycleStatusFailed
-	service.mutateSessionForControl(state, LifecycleControlRetryDispatch, RetryDispatchRequest{DispatchID: "disp-1"}, InterruptDispatchRequest{})
-	if state.session.Status != LifecycleStatusRunning || state.dispatches[0].Status != DispatchStatusQueued || state.dispatches[0].Attempt != 2 {
-		t.Fatalf("retry mutate state = %#v / %#v", state.session, state.dispatches[0])
-	}
-	state.dispatches[0].Status = DispatchStatusRunning
-	state.dispatches[0].Attempt = 1
-	service.mutateSessionForControl(state, LifecycleControlInterruptDispatch, RetryDispatchRequest{}, InterruptDispatchRequest{DispatchID: "disp-1"})
-	if state.dispatches[0].Status != DispatchStatusInterrupted {
-		t.Fatalf("interrupt mutate status = %q, want INTERRUPTED", state.dispatches[0].Status)
-	}
-}
-
-func TestFakeService_InternalStartAndProjectionHelpers(t *testing.T) {
-	t.Parallel()
-	state := newFakeServiceInternalStartState()
-	service := &FakeService{}
-
-	t.Run("start projections", func(t *testing.T) {
-		async := service.asyncStartFromState(state)
-		if async.SessionID != "dur-sess-1" || async.Policy.EffectiveHash != "policy" {
-			t.Fatalf("asyncStartFromState = %#v", async)
-		}
-
-		sync := service.syncStartFromState(state)
-		if sync.SyncOutcome != SyncOutcomeCompleted || len(sync.Result) == 0 {
-			t.Fatalf("syncStartFromState = %#v", sync)
-		}
-		nonTerminal := *state
-		nonTerminal.session.Status = LifecycleStatusRunning
-		sync = service.syncStartFromState(&nonTerminal)
-		if sync.SyncOutcome != "" || len(sync.Result) != 0 {
-			t.Fatalf("non-terminal syncStartFromState = %#v", sync)
-		}
-
-		scenarioAsync := service.asyncStartFromScenario(FakeScenario{AsyncStart: &AsyncStartResult{SessionID: "override"}}, state)
-		if scenarioAsync.SessionID != "override" {
-			t.Fatalf("asyncStartFromScenario = %#v", scenarioAsync)
-		}
-		scenarioSync := service.syncStartFromScenario(FakeScenario{SyncStart: &SyncStartResult{AsyncStartResult: AsyncStartResult{SessionID: "override-sync"}}}, state)
-		if scenarioSync.SessionID != "override-sync" {
-			t.Fatalf("syncStartFromScenario = %#v", scenarioSync)
-		}
-	})
-
-	t.Run("result projections and clones", func(t *testing.T) {
-		testFakeServiceInternalResultProjectionHelpers(t, state, service)
-	})
-
-	t.Run("sync wait outcome", func(t *testing.T) {
-		testFakeServiceInternalSyncWaitHelpers(t, state)
-	})
-}
-
-func newFakeServiceInternalStartState() *fakeSessionState {
-	return &fakeSessionState{
-		session: SessionReadResult{
-			SessionID:        "dur-sess-1",
-			Status:           LifecycleStatusSucceeded,
-			OrchestratorKind: "JAVASCRIPT",
-			Dialect:          "v1",
-			ResolvedSource:   ResolvedSource{Kind: factory.WorkflowSourceKindWorkflowName, SourceRef: "audit", SourceHash: "hash", Dialect: "v1"},
-			SourceHash:       "hash",
-			Policy:           PolicyProjection{EffectiveHash: "policy"},
-			Links:            InspectionLinks{Session: "/factory-sessions/dur-sess-1"},
-		},
-		result: ResultReadResult{
-			SessionID:        "dur-sess-1",
-			ResultStatus:     ResultStatusFinal,
-			SessionStatus:    LifecycleStatusSucceeded,
-			PrimaryResult:    json.RawMessage(`[{"type":"text","text":"done"}]`),
-			Availability:     &ResultAvailabilityDetail{Reason: "IGNORED"},
-			IncludeArtifacts: true,
-		},
-		artifacts: []ArtifactSummary{
-			{ID: "art-1", Kind: "LOG", Visibility: "PUBLIC", ContentHash: "hash-1", SizeBytes: 7},
-			{ID: " ", Kind: "LOG", Visibility: "PUBLIC"},
-		},
-		events: []json.RawMessage{json.RawMessage(`{"id":"event-1"}`)},
-	}
-}
-
-func testFakeServiceInternalResultProjectionHelpers(t *testing.T, state *fakeSessionState, service *FakeService) {
-	t.Helper()
-
-	async := service.asyncStartFromState(state)
-	sync := service.syncStartFromState(state)
-
-	t.Run("projection modes", func(t *testing.T) {
-		testFakeServiceInternalResultProjectionModes(t, state)
-	})
-	t.Run("helper branches and clones", func(t *testing.T) {
-		testFakeServiceInternalResultHelperBranches(t, state, async, sync)
-	})
-}
-
-func testFakeServiceInternalResultProjectionModes(t *testing.T, state *fakeSessionState) {
-	t.Helper()
-
-	canonical := ResultReadResult{
-		SessionID:     "dur-sess-1",
-		ResultStatus:  ResultStatusPartial,
-		PrimaryResult: json.RawMessage(`[{"type":"text","text":"partial"}]`),
-		Failure:       &FailureSummary{Reason: "warn", PartialResultAvailable: true},
-	}
-	session := SessionReadResult{
-		SessionID: "dur-sess-1",
-		Status:    LifecycleStatusRunning,
-		ResultSummary: &ResultSummary{
-			ResultStatus: string(ResultStatusPartial),
-		},
-	}
-	projected, err := ProjectResultRead(canonical, session, state.artifacts, ResultRequest{Mode: ResultModeFinal, IncludeArtifacts: true})
-	if err != nil {
-		t.Fatalf("ProjectResultRead final: %v", err)
-	}
-	if projected.ResultStatus != ResultStatusNotReady || projected.Availability == nil || len(projected.ArtifactRefs) != 2 {
-		t.Fatalf("projected final = %#v", projected)
-	}
-
-	projected, err = ProjectResultRead(canonical, session, state.artifacts, ResultRequest{Mode: ResultModePartial})
-	if err != nil {
-		t.Fatalf("ProjectResultRead partial: %v", err)
-	}
-	if projected.ResultStatus != ResultStatusPartial || len(projected.PrimaryResult) == 0 || projected.Availability != nil {
-		t.Fatalf("projected partial = %#v", projected)
-	}
-	if _, err := ProjectResultRead(canonical, session, nil, ResultRequest{Mode: ResultMode("bad")}); err == nil {
-		t.Fatal("invalid mode should fail normalization")
-	}
-}
-
-func testFakeServiceInternalResultHelperBranches(t *testing.T, state *fakeSessionState, async AsyncStartResult, sync SyncStartResult) {
-	t.Helper()
-
-	t.Run("status and artifact helpers", func(t *testing.T) {
-		testFakeServiceInternalResultStatusAndArtifactHelpers(t, state)
-	})
-	t.Run("clone helpers", func(t *testing.T) {
-		testFakeServiceInternalResultCloneHelpers(t, async, sync)
-	})
-}
-
-func testFakeServiceInternalResultStatusAndArtifactHelpers(t *testing.T, state *fakeSessionState) {
-	t.Helper()
-
-	if got := canonicalResultStatus(ResultReadResult{ResultStatus: ResultStatusUnavailable}, SessionReadResult{
-		ResultSummary: &ResultSummary{ResultStatus: " FINAL "},
-	}); got != ResultStatusFinal {
-		t.Fatalf("canonicalResultStatus = %q", got)
-	}
-	if got := defaultNotReadyAvailability(SessionReadResult{Status: LifecycleStatusRunning}); got == nil || !got.Retryable {
-		t.Fatalf("running default availability = %#v", got)
-	}
-	if got := defaultNotReadyAvailability(SessionReadResult{Status: LifecycleStatusSucceeded}); got == nil || got.Retryable {
-		t.Fatalf("terminal default availability = %#v", got)
-	}
-
-	if refs := artifactRefsFromSummaries(nil); refs != nil {
-		t.Fatalf("artifactRefsFromSummaries(nil) = %#v", refs)
-	}
-	refs := artifactRefsFromSummaries(state.artifacts)
-	if len(refs) != 2 || refs[0].ID != "art-1" {
-		t.Fatalf("artifactRefsFromSummaries = %#v", refs)
-	}
-	ids := artifactIDsFromSummaries(state.artifacts)
-	if len(ids) != 1 || ids[0] != "art-1" {
-		t.Fatalf("artifactIDsFromSummaries = %#v", ids)
-	}
-}
-
-func testFakeServiceInternalResultCloneHelpers(t *testing.T, async AsyncStartResult, sync SyncStartResult) {
-	t.Helper()
-
-	canonicalFailure := &FailureSummary{Reason: "warn", PartialResultAvailable: true}
-
-	if cloneFailureSummary(nil) != nil || cloneResultAvailability(nil) != nil || cloneRawJSON(nil) != nil {
-		t.Fatal("nil clones should stay nil")
-	}
-	if clone := cloneFailureSummary(canonicalFailure); clone == canonicalFailure || clone.Reason != "warn" {
-		t.Fatalf("cloneFailureSummary = %#v", clone)
-	}
-	if clone := cloneResultAvailability(&ResultAvailabilityDetail{Reason: "NOT_READY"}); clone == nil || clone.Reason != "NOT_READY" {
-		t.Fatalf("cloneResultAvailability = %#v", clone)
-	}
-	if clone := cloneAsyncStartResult(async); clone.SessionID != async.SessionID {
-		t.Fatalf("cloneAsyncStartResult = %#v", clone)
-	}
-	if clone := cloneSyncStartResult(sync); clone.SessionID != sync.SessionID {
-		t.Fatalf("cloneSyncStartResult = %#v", clone)
-	}
-	if clone := cloneRawJSON(json.RawMessage(`{"ok":true}`)); string(clone) != `{"ok":true}` {
-		t.Fatalf("cloneRawJSON = %s", clone)
-	}
-}
-
-func testFakeServiceInternalSyncWaitHelpers(t *testing.T, state *fakeSessionState) {
-	t.Helper()
-
-	applySyncWaitOutcome(nil, state, StartRequest{})
-	applySyncWaitOutcome(&SyncStartResult{}, nil, StartRequest{})
-	timeout := SyncStartResult{AsyncStartResult: AsyncStartResult{Status: string(LifecycleStatusRunning)}, SyncOutcome: SyncOutcomeTimedOut, TimedOut: true}
-	runningState := &fakeSessionState{
-		session: SessionReadResult{SessionID: "dur-sess-2", Status: LifecycleStatusRunning},
-		result:  ResultReadResult{SessionID: "dur-sess-2", SessionStatus: LifecycleStatusRunning, ResultStatus: ResultStatusNotReady},
-	}
-	applySyncWaitOutcome(&timeout, runningState, StartRequest{
-		Wait: &WaitOptions{CancelOnTimeout: true},
-	})
-	if !timeout.SessionCanceledByTimeout || timeout.Status != string(LifecycleStatusCanceling) || runningState.result.Availability == nil {
-		t.Fatalf("applySyncWaitOutcome timeout = %#v / %#v", timeout, runningState.result)
 	}
 }
 
@@ -1860,54 +1587,6 @@ func writeSimpleFinalWorkflowProject(t *testing.T) string {
 		t.Fatalf("write workflow: %v", err)
 	}
 	return projectRoot
-}
-
-func TestFakeService_DetailReadersAndRemainingControlWrappers(t *testing.T) {
-	t.Parallel()
-	service := newContractFakeService(t)
-
-	startAsyncByRequestID(t, service, "req-petri-success-001")
-	dispatch, err := service.GetDispatch(context.Background(), "dur-sess-petri-success-001", "disp-petri-success-001")
-	if err != nil {
-		t.Fatalf("GetDispatch: %v", err)
-	}
-	if dispatch.DispatchSummary.ID != "disp-petri-success-001" {
-		t.Fatalf("dispatch detail = %#v", dispatch)
-	}
-
-	artifact, err := service.GetArtifact(context.Background(), "dur-sess-petri-success-001", "art-petri-final-001")
-	if err != nil {
-		t.Fatalf("GetArtifact: %v", err)
-	}
-	if artifact.ArtifactSummary.ID != "art-petri-final-001" {
-		t.Fatalf("artifact detail = %#v", artifact)
-	}
-
-	startAsyncByRequestID(t, service, "req-js-run-n-001")
-	cancelled, err := service.Cancel(context.Background(), "dur-sess-js-run-n-001", ControlRequest{})
-	if err != nil {
-		t.Fatalf("Cancel: %v", err)
-	}
-	if cancelled.Status != LifecycleStatusCanceling {
-		t.Fatalf("cancel status = %q, want CANCELING", cancelled.Status)
-	}
-
-	terminated, err := service.Terminate(context.Background(), "dur-sess-js-run-n-001", ControlRequest{})
-	if err != nil {
-		t.Fatalf("Terminate: %v", err)
-	}
-	if terminated.Status != LifecycleStatusTerminated {
-		t.Fatalf("terminate status = %q, want TERMINATED", terminated.Status)
-	}
-
-	startAsyncByRequestID(t, service, "req-js-awaiting-001")
-	approved, err := service.Approve(context.Background(), "dur-sess-js-awaiting-001", ApproveRequest{})
-	if err != nil {
-		t.Fatalf("Approve: %v", err)
-	}
-	if approved.Status != LifecycleStatusRunning {
-		t.Fatalf("approve status = %q, want RUNNING", approved.Status)
-	}
 }
 
 func TestJavaScriptRuntimeService_ControlWrappersAndDetailReaders(t *testing.T) {
